parameter maintenance on admin ui, demo content

+++ src/main/java/com/opencommunity/openTeamOneServer/ContentService.java
@@ -22,19 +22,14 @@ public class ContentService {
 		return null;
 	}
 
-	public static void importFromJson(JSONObject jsonObject, boolean includeConfiguration, boolean delete, String protectedUserId) throws JSONException {
+	public static void importFromJson(JSONObject jsonObject, boolean delete, boolean includeConfiguration, String protectedUserId) throws JSONException {
 		if (instance != null)
-			instance._importFromJson(jsonObject, includeConfiguration, delete, protectedUserId);
+			instance._importFromJson(jsonObject, delete, includeConfiguration, protectedUserId);
 	}
 
-	public static void deleteAll(String protectedUserId) throws JSONException {
+	public static void deleteAll(boolean includeConfiguration, String protectedUserId) throws JSONException {
 		if (instance != null)
-			instance._deleteAll(protectedUserId);
-	}
-
-	public static void loadModelData() throws JSONException {
-		if (instance != null)
-			instance._loadModelData();
+			instance._deleteAll(includeConfiguration, protectedUserId);
 	}
 
 	/* instance methods and properties */
@@ -67,7 +62,7 @@ public class ContentService {
 		if (tpr.findOne("name") == null)
 			tpr.save(new TenantParameter("name", "OpenTeamOne"));
 		if (tpr.findOne("pictureId") == null)
-			tpr.save(new TenantParameter("pictureId", "logo"));
+			tpr.save(new TenantParameter("pictureId", "tenant"));
 		if (tpr.findOne("startPageNoLogon") == null)
 			tpr.save(new TenantParameter("startPageNoLogon", "/admin/default/index.html"));
 		if (tpr.findOne("startPageNoAdmin") == null)
@@ -95,7 +90,7 @@ public class ContentService {
 		return jsonObject;
 	}
 
-	private void _importFromJson(JSONObject jsonObject, boolean includeConfiguration, boolean delete, String protectedUserId) throws JSONException {
+	private void _importFromJson(JSONObject jsonObject, boolean delete, boolean includeConfiguration, String protectedUserId) throws JSONException {
 		JSONArray item;
 		item = JsonUtil.getJSONArray(jsonObject, "tenantParameters");
 		if (item != null && includeConfiguration) {
@@ -163,77 +158,17 @@ public class ContentService {
 		}
 	}
 
-	private void _deleteAll(String protectedUserId) {
-		tpr.deleteAll();
+	private void _deleteAll(boolean includeConfiguration, String protectedUserId) {
+		if (includeConfiguration)
+			tpr.deleteAll();
 		if (protectedUserId != null)
-			ur.delete(ur.findByUserIdNot(protectedUserId));
+			ur.delete(includeConfiguration ? ur.findByUserIdNot(protectedUserId) : ur.findByHasAdminRoleFalseAndUserIdNot(protectedUserId));
 		pr.deleteAll();
 		rr.deleteAll();
 		rmr.deleteAll();
 		mr.deleteAll();
 		ar.deleteAll();
 		vcr.deleteAll();
-	}
-
-	private void _loadModelData() {
-		long now = System.currentTimeMillis();
-		//
-		tpr.save(new TenantParameter("name", "OpenTeamOne"));
-		tpr.save(new TenantParameter("pictureId", "tenant"));
-		tpr.save(new TenantParameter("dataDirectory", "/var/cache/openTeamOne"));
-		//
-		Person p1, p2, p3;
-		pr.save(p1 = new Person(null, "Tank", "Thomas", "Tom", "profile1"));
-		pr.save(p2 = new Person(null, "Smith", "Peter", null, "profile2"));
-		pr.save(p3 = new Person(null, "Potter", "Harry", null, "profile3"));
-		//
-		ur.save(new User("player01", "pass", p1.personId, true, false));
-		ur.save(new User("player02", "pass", p2.personId, true, false));
-		ur.save(new User("player03", "pass", p3.personId, true, false));
-		//
-		Room r1, r2, r3;
-		rr.save(r1 = new Room(null, "Team Room", "TR", "group", "room1", now - 100000L));
-		rr.save(r2 = new Room(null, "General", "GN", "group", "room2", now - 200000L));
-		rr.save(r3 = new Room(null, "PM", "PM", "private", null, now - 300000L));
-		//
-		rmr.save(new RoomMember(r1.roomId, p1.personId));
-		rmr.save(new RoomMember(r1.roomId, p2.personId));
-		rmr.save(new RoomMember(r1.roomId, p3.personId));
-		rmr.save(new RoomMember(r2.roomId, p1.personId));
-		rmr.save(new RoomMember(r2.roomId, p2.personId));
-		rmr.save(new RoomMember(r2.roomId, p3.personId));
-		rmr.save(new RoomMember(r3.roomId, p1.personId));
-		rmr.save(new RoomMember(r3.roomId, p2.personId));
-		//
-		Message m1, m2, m3, m4, m5, m6;
-		mr.save(m1 = new Message(null, "cm1", r1.roomId, p1.personId, now - 500000, "Hello, this is my first message", false, now - 500000));
-		mr.save(m2 = new Message(null, "cm2", r1.roomId, p3.personId, now - 100000, "Welcome! \uD83D\uDE0A", false, now - 100000));
-		mr.save(m3 = new Message(null, "cm3", r2.roomId, p2.personId, now - 300000, "Can you see my picture?", false, now - 300000));
-		mr.save(m4 = new Message(null, "cm4", r2.roomId, p1.personId, now - 200000, "Yeah, brilliant shot! \uD83D\uDC4D", false, now - 200000));
-		mr.save(m5 = new Message(null, "cm5", r3.roomId, p1.personId, now - 600000, "Hi Peter", false, now - 600000));
-		mr.save(m6 = new Message(null, "cm6", r3.roomId, p2.personId, now - 300000, "Hey Tom, everything allright?", false, now - 300000));
-		//
-		ar.save(new Attachment("tenant", "image/jpg", null, null));
-		ar.save(new Attachment("profile1", "image/jpg", null, null));
-		ar.save(new Attachment("profile2", "image/jpg", null, null));
-		ar.save(new Attachment("profile3", "image/jpg", null, null));
-		ar.save(new Attachment("room1", "image/jpg", null, null));
-		ar.save(new Attachment("room2", "image/jpg", null, null));
-		ar.save(new Attachment("sailing1", "image/jpg", null, m3.messageId));
-		//
-		vcr.save(new ViewedConfirmation(m1.messageId, p1.personId, m1.roomId, m1.postedAt, now));
-		vcr.save(new ViewedConfirmation(m1.messageId, p2.personId, m1.roomId, m1.postedAt, now));
-		vcr.save(new ViewedConfirmation(m1.messageId, p3.personId, m1.roomId, m1.postedAt, now));
-		vcr.save(new ViewedConfirmation(m2.messageId, p1.personId, m2.roomId, m2.postedAt, now));
-		vcr.save(new ViewedConfirmation(m2.messageId, p3.personId, m2.roomId, m2.postedAt, now));
-		vcr.save(new ViewedConfirmation(m3.messageId, p1.personId, m3.roomId, m3.postedAt, now));
-		vcr.save(new ViewedConfirmation(m3.messageId, p2.personId, m3.roomId, m3.postedAt, now));
-		vcr.save(new ViewedConfirmation(m3.messageId, p3.personId, m3.roomId, m3.postedAt, now));
-		vcr.save(new ViewedConfirmation(m4.messageId, p1.personId, m4.roomId, m4.postedAt, now));
-		vcr.save(new ViewedConfirmation(m4.messageId, p2.personId, m4.roomId, m4.postedAt, now));
-		vcr.save(new ViewedConfirmation(m5.messageId, p1.personId, m5.roomId, m5.postedAt, now));
-		vcr.save(new ViewedConfirmation(m5.messageId, p2.personId, m5.roomId, m5.postedAt, now));
-		vcr.save(new ViewedConfirmation(m6.messageId, p2.personId, m6.roomId, m6.postedAt, now));
 	}
 
 }
